adding http auth

+++ app/templates/handler/plural.go
@@ -7,12 +7,20 @@ import (
     "io/ioutil"
     "encoding/json"
 
+    <%if (auth) { %>"<%= vcs %>/<%= repo %>/<%= project %>/auth"<% } %>
     "<%= vcs %>/<%= repo %>/<%= project %>/utils"
     "<%= vcs %>/<%= repo %>/<%= project %>/models"
     "<%= vcs %>/<%= repo %>/<%= project %>/database"
 )
 
 func Retrieve<%= nounPluralUpper %>(db *database.Database, w http.ResponseWriter, r *http.Request) error {
+
+    <%if (auth) { %>
+    if(!auth.Check(w, r)) {
+        return utils.StatusError{http.StatusUnauthorized, errors.New("unauthorized")}
+    }
+    <% } %>
+
     result, err := db.RetrieveAll()
     if err != nil {
         return utils.StatusError{http.StatusInternalServerError, err}
@@ -30,6 +38,13 @@ func Retrieve<%= nounPluralUpper %>(db *database.Database, w http.ResponseWriter
 }
 
 func Create<%= nounSingularUpper %>(db *database.Database, w http.ResponseWriter, r *http.Request) error {
+
+    <%if (auth) { %>
+    if(!auth.Check(w, r)) {
+        return utils.StatusError{http.StatusUnauthorized, errors.New("unauthorized")}
+    }
+    <% } %>
+
     <%= nounSingularLower %> := models.<%= nounSingularUpper %>{}
 	body, err := ioutil.ReadAll(io.LimitReader(r.Body, 1048576))
     if err != nil {
@@ -47,7 +62,7 @@ func Create<%= nounSingularUpper %>(db *database.Database, w http.ResponseWriter
     }
     w.Header().Set("location", fmt.Sprintf("/<%= nounSingularLower %>/%s", result.ID))
     w.Header().Set("Content-Type", "application/json")
-    w.WriteHeader(http.StatusOK)
+    w.WriteHeader(http.StatusCreated)
     if err := json.NewEncoder(w).Encode(result); err != nil {
         return utils.StatusError{http.StatusInternalServerError, err}
     }
@@ -55,6 +70,13 @@ func Create<%= nounSingularUpper %>(db *database.Database, w http.ResponseWriter
 }
 
 func Delete<%= nounPluralUpper %>(db *database.Database, w http.ResponseWriter, r *http.Request) error {
+
+    <%if (auth) { %>
+    if(!auth.Check(w, r)) {
+        return utils.StatusError{http.StatusUnauthorized, errors.New("unauthorized")}
+    }
+    <% } %>
+
     if _, err := db.DeleteAll(); err != nil {
         return utils.StatusError{http.StatusInternalServerError, err}
     }
