+++ app/templates/handler/singular.go
@@ -7,12 +7,20 @@ import (
     "encoding/json"
 
     "github.com/gorilla/mux"
+    <%if (auth) { %>"<%= vcs %>/<%= repo %>/<%= project %>/auth"<% } %>
     "<%= vcs %>/<%= repo %>/<%= project %>/utils"
     "<%= vcs %>/<%= repo %>/<%= project %>/models"
     "<%= vcs %>/<%= repo %>/<%= project %>/database"
 )
 
 func Retrieve<%= nounSingularUpper %>(db *database.Database, w http.ResponseWriter, r *http.Request) error {
+
+    <%if (auth) { %>
+    if(!auth.Check(w, r)) {
+        return utils.StatusError{http.StatusUnauthorized, errors.New("unauthorized")}
+    }
+    <% } %>
+
     id := mux.Vars(r)["id"]
     <%= nounSingularLower %>, err := db.RetrieveOne(id)
     if err != nil {
@@ -31,6 +39,13 @@ func Retrieve<%= nounSingularUpper %>(db *database.Database, w http.ResponseWrit
 }
 
 func Update<%= nounSingularUpper %>(db *database.Database, w http.ResponseWriter, r *http.Request) error {
+
+    <%if (auth) { %>
+    if(!auth.Check(w, r)) {
+        return utils.StatusError{http.StatusUnauthorized, errors.New("unauthorized")}
+    }
+    <% } %>
+
     id := mux.Vars(r)["id"]
     <%= nounSingularLower %> := models.<%= nounSingularUpper %>{}
 	body, err := ioutil.ReadAll(io.LimitReader(r.Body, 1048576))
@@ -60,6 +75,13 @@ func Update<%= nounSingularUpper %>(db *database.Database, w http.ResponseWriter
 }
 
 func Delete<%= nounSingularUpper %>(db *database.Database, w http.ResponseWriter, r *http.Request) error {
+
+    <%if (auth) { %>
+    if(!auth.Check(w, r)) {
+        return utils.StatusError{http.StatusUnauthorized, errors.New("unauthorized")}
+    }
+    <% } %>
+
     id := mux.Vars(r)["id"]
     result, err := db.DeleteOne(id)
     if err != nil {
